+++ apm-agent-plugins/apm-spring-webflux/apm-spring-webflux-spring5/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.internal.PluginClassLoaderRootPackageCustomizer
@@ -1,0 +1,1 @@
+co.elastic.apm.agent.springwebflux.WebfluxPluginRootPackageCustomizer
